+++ functional/otlp/func_grpc_main.cc
@@ -56,6 +56,9 @@ static std::string opt_cert_dir;
 static std::string opt_test_name;
 static TestMode opt_mode = TestMode::kNone;
 
+namespace
+{
+
 /*
   Log parsing
 */
@@ -318,6 +321,8 @@ struct test_case
   std::string m_name;
   test_func_t m_func;
 };
+
+}  // namespace
 
 static int test_basic();