+++ CHANGELOG.md
@@ -51,6 +51,9 @@ Increment the:
 * [CODE HEALTH] Move func_http_main classes into anonymous namespace
   [#4128](https://github.com/open-telemetry/opentelemetry-cpp/pull/4128)
 
+* [CODE HEALTH] Move func_grpc_main classes into anonymous namespace
+  [#4129](https://github.com/open-telemetry/opentelemetry-cpp/pull/4129)
+
 ## [1.27.0] 2026-05-13
 
 * [RELEASE] Bump main branch to 1.27.0-dev